requirements.txt
streamlit
pandas
streamlit-aggrid
streamlit
opencv-python
Pillow
numpy
pandas